backend/requirements.txt
asgiref==3.9.1
black==25.9.0
boto3==1.40.41
botocore==1.40.41
click==8.3.0
colorama==0.4.6
django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
flake8==7.3.0
isort==6.0.1
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytokens==0.1.10
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
